Create utility functions to fill gas maps; relocate WASM-related struct method
