--- config/wasmGasCost.go
@@ -1,835 +1,0 @@
-package config
-
-import "github.com/ElrondNetwork/go-ext-wasm/wasmer"
-
-func FillGasMapWithWASMOpcodeValues(gasMap map[string]uint64, value uint64) map[string]uint64 {
-	gasMap["Unreachable"] = value
-	gasMap["Nop"] = value
-	gasMap["Block"] = value
-	gasMap["Loop"] = value
-	gasMap["If"] = value
-	gasMap["Else"] = value
-	gasMap["End"] = value
-	gasMap["Br"] = value
-	gasMap["BrIf"] = value
-	gasMap["BrTable"] = value
-	gasMap["Return"] = value
-	gasMap["Call"] = value
-	gasMap["CallIndirect"] = value
-	gasMap["Drop"] = value
-	gasMap["Select"] = value
-	gasMap["GetLocal"] = value
-	gasMap["SetLocal"] = value
-	gasMap["TeeLocal"] = value
-	gasMap["GetGlobal"] = value
-	gasMap["SetGlobal"] = value
-	gasMap["I32Load"] = value
-	gasMap["I64Load"] = value
-	gasMap["F32Load"] = value
-	gasMap["F64Load"] = value
-	gasMap["I32Load8S"] = value
-	gasMap["I32Load8U"] = value
-	gasMap["I32Load16S"] = value
-	gasMap["I32Load16U"] = value
-	gasMap["I64Load8S"] = value
-	gasMap["I64Load8U"] = value
-	gasMap["I64Load16S"] = value
-	gasMap["I64Load16U"] = value
-	gasMap["I64Load32S"] = value
-	gasMap["I64Load32U"] = value
-	gasMap["I32Store"] = value
-	gasMap["I64Store"] = value
-	gasMap["F32Store"] = value
-	gasMap["F64Store"] = value
-	gasMap["I32Store8"] = value
-	gasMap["I32Store16"] = value
-	gasMap["I64Store8"] = value
-	gasMap["I64Store16"] = value
-	gasMap["I64Store32"] = value
-	gasMap["MemorySize"] = value
-	gasMap["MemoryGrow"] = value
-	gasMap["I32Const"] = value
-	gasMap["I64Const"] = value
-	gasMap["F32Const"] = value
-	gasMap["F64Const"] = value
-	gasMap["RefNull"] = value
-	gasMap["RefIsNull"] = value
-	gasMap["I32Eqz"] = value
-	gasMap["I32Eq"] = value
-	gasMap["I32Ne"] = value
-	gasMap["I32LtS"] = value
-	gasMap["I32LtU"] = value
-	gasMap["I32GtS"] = value
-	gasMap["I32GtU"] = value
-	gasMap["I32LeS"] = value
-	gasMap["I32LeU"] = value
-	gasMap["I32GeS"] = value
-	gasMap["I32GeU"] = value
-	gasMap["I64Eqz"] = value
-	gasMap["I64Eq"] = value
-	gasMap["I64Ne"] = value
-	gasMap["I64LtS"] = value
-	gasMap["I64LtU"] = value
-	gasMap["I64GtS"] = value
-	gasMap["I64GtU"] = value
-	gasMap["I64LeS"] = value
-	gasMap["I64LeU"] = value
-	gasMap["I64GeS"] = value
-	gasMap["I64GeU"] = value
-	gasMap["F32Eq"] = value
-	gasMap["F32Ne"] = value
-	gasMap["F32Lt"] = value
-	gasMap["F32Gt"] = value
-	gasMap["F32Le"] = value
-	gasMap["F32Ge"] = value
-	gasMap["F64Eq"] = value
-	gasMap["F64Ne"] = value
-	gasMap["F64Lt"] = value
-	gasMap["F64Gt"] = value
-	gasMap["F64Le"] = value
-	gasMap["F64Ge"] = value
-	gasMap["I32Clz"] = value
-	gasMap["I32Ctz"] = value
-	gasMap["I32Popcnt"] = value
-	gasMap["I32Add"] = value
-	gasMap["I32Sub"] = value
-	gasMap["I32Mul"] = value
-	gasMap["I32DivS"] = value
-	gasMap["I32DivU"] = value
-	gasMap["I32RemS"] = value
-	gasMap["I32RemU"] = value
-	gasMap["I32And"] = value
-	gasMap["I32Or"] = value
-	gasMap["I32Xor"] = value
-	gasMap["I32Shl"] = value
-	gasMap["I32ShrS"] = value
-	gasMap["I32ShrU"] = value
-	gasMap["I32Rotl"] = value
-	gasMap["I32Rotr"] = value
-	gasMap["I64Clz"] = value
-	gasMap["I64Ctz"] = value
-	gasMap["I64Popcnt"] = value
-	gasMap["I64Add"] = value
-	gasMap["I64Sub"] = value
-	gasMap["I64Mul"] = value
-	gasMap["I64DivS"] = value
-	gasMap["I64DivU"] = value
-	gasMap["I64RemS"] = value
-	gasMap["I64RemU"] = value
-	gasMap["I64And"] = value
-	gasMap["I64Or"] = value
-	gasMap["I64Xor"] = value
-	gasMap["I64Shl"] = value
-	gasMap["I64ShrS"] = value
-	gasMap["I64ShrU"] = value
-	gasMap["I64Rotl"] = value
-	gasMap["I64Rotr"] = value
-	gasMap["F32Abs"] = value
-	gasMap["F32Neg"] = value
-	gasMap["F32Ceil"] = value
-	gasMap["F32Floor"] = value
-	gasMap["F32Trunc"] = value
-	gasMap["F32Nearest"] = value
-	gasMap["F32Sqrt"] = value
-	gasMap["F32Add"] = value
-	gasMap["F32Sub"] = value
-	gasMap["F32Mul"] = value
-	gasMap["F32Div"] = value
-	gasMap["F32Min"] = value
-	gasMap["F32Max"] = value
-	gasMap["F32Copysign"] = value
-	gasMap["F64Abs"] = value
-	gasMap["F64Neg"] = value
-	gasMap["F64Ceil"] = value
-	gasMap["F64Floor"] = value
-	gasMap["F64Trunc"] = value
-	gasMap["F64Nearest"] = value
-	gasMap["F64Sqrt"] = value
-	gasMap["F64Add"] = value
-	gasMap["F64Sub"] = value
-	gasMap["F64Mul"] = value
-	gasMap["F64Div"] = value
-	gasMap["F64Min"] = value
-	gasMap["F64Max"] = value
-	gasMap["F64Copysign"] = value
-	gasMap["I32WrapI64"] = value
-	gasMap["I32TruncSF32"] = value
-	gasMap["I32TruncUF32"] = value
-	gasMap["I32TruncSF64"] = value
-	gasMap["I32TruncUF64"] = value
-	gasMap["I64ExtendSI32"] = value
-	gasMap["I64ExtendUI32"] = value
-	gasMap["I64TruncSF32"] = value
-	gasMap["I64TruncUF32"] = value
-	gasMap["I64TruncSF64"] = value
-	gasMap["I64TruncUF64"] = value
-	gasMap["F32ConvertSI32"] = value
-	gasMap["F32ConvertUI32"] = value
-	gasMap["F32ConvertSI64"] = value
-	gasMap["F32ConvertUI64"] = value
-	gasMap["F32DemoteF64"] = value
-	gasMap["F64ConvertSI32"] = value
-	gasMap["F64ConvertUI32"] = value
-	gasMap["F64ConvertSI64"] = value
-	gasMap["F64ConvertUI64"] = value
-	gasMap["F64PromoteF32"] = value
-	gasMap["I32ReinterpretF32"] = value
-	gasMap["I64ReinterpretF64"] = value
-	gasMap["F32ReinterpretI32"] = value
-	gasMap["F64ReinterpretI64"] = value
-	gasMap["I32Extend8S"] = value
-	gasMap["I32Extend16S"] = value
-	gasMap["I64Extend8S"] = value
-	gasMap["I64Extend16S"] = value
-	gasMap["I64Extend32S"] = value
-	gasMap["I32TruncSSatF32"] = value
-	gasMap["I32TruncUSatF32"] = value
-	gasMap["I32TruncSSatF64"] = value
-	gasMap["I32TruncUSatF64"] = value
-	gasMap["I64TruncSSatF32"] = value
-	gasMap["I64TruncUSatF32"] = value
-	gasMap["I64TruncSSatF64"] = value
-	gasMap["I64TruncUSatF64"] = value
-	gasMap["MemoryInit"] = value
-	gasMap["DataDrop"] = value
-	gasMap["MemoryCopy"] = value
-	gasMap["MemoryFill"] = value
-	gasMap["TableInit"] = value
-	gasMap["ElemDrop"] = value
-	gasMap["TableCopy"] = value
-	gasMap["TableGet"] = value
-	gasMap["TableSet"] = value
-	gasMap["TableGrow"] = value
-	gasMap["TableSize"] = value
-	gasMap["Wake"] = value
-	gasMap["I32Wait"] = value
-	gasMap["I64Wait"] = value
-	gasMap["Fence"] = value
-	gasMap["I32AtomicLoad"] = value
-	gasMap["I64AtomicLoad"] = value
-	gasMap["I32AtomicLoad8U"] = value
-	gasMap["I32AtomicLoad16U"] = value
-	gasMap["I64AtomicLoad8U"] = value
-	gasMap["I64AtomicLoad16U"] = value
-	gasMap["I64AtomicLoad32U"] = value
-	gasMap["I32AtomicStore"] = value
-	gasMap["I64AtomicStore"] = value
-	gasMap["I32AtomicStore8"] = value
-	gasMap["I32AtomicStore16"] = value
-	gasMap["I64AtomicStore8"] = value
-	gasMap["I64AtomicStore16"] = value
-	gasMap["I64AtomicStore32"] = value
-	gasMap["I32AtomicRmwAdd"] = value
-	gasMap["I64AtomicRmwAdd"] = value
-	gasMap["I32AtomicRmw8UAdd"] = value
-	gasMap["I32AtomicRmw16UAdd"] = value
-	gasMap["I64AtomicRmw8UAdd"] = value
-	gasMap["I64AtomicRmw16UAdd"] = value
-	gasMap["I64AtomicRmw32UAdd"] = value
-	gasMap["I32AtomicRmwSub"] = value
-	gasMap["I64AtomicRmwSub"] = value
-	gasMap["I32AtomicRmw8USub"] = value
-	gasMap["I32AtomicRmw16USub"] = value
-	gasMap["I64AtomicRmw8USub"] = value
-	gasMap["I64AtomicRmw16USub"] = value
-	gasMap["I64AtomicRmw32USub"] = value
-	gasMap["I32AtomicRmwAnd"] = value
-	gasMap["I64AtomicRmwAnd"] = value
-	gasMap["I32AtomicRmw8UAnd"] = value
-	gasMap["I32AtomicRmw16UAnd"] = value
-	gasMap["I64AtomicRmw8UAnd"] = value
-	gasMap["I64AtomicRmw16UAnd"] = value
-	gasMap["I64AtomicRmw32UAnd"] = value
-	gasMap["I32AtomicRmwOr"] = value
-	gasMap["I64AtomicRmwOr"] = value
-	gasMap["I32AtomicRmw8UOr"] = value
-	gasMap["I32AtomicRmw16UOr"] = value
-	gasMap["I64AtomicRmw8UOr"] = value
-	gasMap["I64AtomicRmw16UOr"] = value
-	gasMap["I64AtomicRmw32UOr"] = value
-	gasMap["I32AtomicRmwXor"] = value
-	gasMap["I64AtomicRmwXor"] = value
-	gasMap["I32AtomicRmw8UXor"] = value
-	gasMap["I32AtomicRmw16UXor"] = value
-	gasMap["I64AtomicRmw8UXor"] = value
-	gasMap["I64AtomicRmw16UXor"] = value
-	gasMap["I64AtomicRmw32UXor"] = value
-	gasMap["I32AtomicRmwXchg"] = value
-	gasMap["I64AtomicRmwXchg"] = value
-	gasMap["I32AtomicRmw8UXchg"] = value
-	gasMap["I32AtomicRmw16UXchg"] = value
-	gasMap["I64AtomicRmw8UXchg"] = value
-	gasMap["I64AtomicRmw16UXchg"] = value
-	gasMap["I64AtomicRmw32UXchg"] = value
-	gasMap["I32AtomicRmwCmpxchg"] = value
-	gasMap["I64AtomicRmwCmpxchg"] = value
-	gasMap["I32AtomicRmw8UCmpxchg"] = value
-	gasMap["I32AtomicRmw16UCmpxchg"] = value
-	gasMap["I64AtomicRmw8UCmpxchg"] = value
-	gasMap["I64AtomicRmw16UCmpxchg"] = value
-	gasMap["I64AtomicRmw32UCmpxchg"] = value
-	gasMap["V128Load"] = value
-	gasMap["V128Store"] = value
-	gasMap["V128Const"] = value
-	gasMap["I8x16Splat"] = value
-	gasMap["I8x16ExtractLaneS"] = value
-	gasMap["I8x16ExtractLaneU"] = value
-	gasMap["I8x16ReplaceLane"] = value
-	gasMap["I16x8Splat"] = value
-	gasMap["I16x8ExtractLaneS"] = value
-	gasMap["I16x8ExtractLaneU"] = value
-	gasMap["I16x8ReplaceLane"] = value
-	gasMap["I32x4Splat"] = value
-	gasMap["I32x4ExtractLane"] = value
-	gasMap["I32x4ReplaceLane"] = value
-	gasMap["I64x2Splat"] = value
-	gasMap["I64x2ExtractLane"] = value
-	gasMap["I64x2ReplaceLane"] = value
-	gasMap["F32x4Splat"] = value
-	gasMap["F32x4ExtractLane"] = value
-	gasMap["F32x4ReplaceLane"] = value
-	gasMap["F64x2Splat"] = value
-	gasMap["F64x2ExtractLane"] = value
-	gasMap["F64x2ReplaceLane"] = value
-	gasMap["I8x16Eq"] = value
-	gasMap["I8x16Ne"] = value
-	gasMap["I8x16LtS"] = value
-	gasMap["I8x16LtU"] = value
-	gasMap["I8x16GtS"] = value
-	gasMap["I8x16GtU"] = value
-	gasMap["I8x16LeS"] = value
-	gasMap["I8x16LeU"] = value
-	gasMap["I8x16GeS"] = value
-	gasMap["I8x16GeU"] = value
-	gasMap["I16x8Eq"] = value
-	gasMap["I16x8Ne"] = value
-	gasMap["I16x8LtS"] = value
-	gasMap["I16x8LtU"] = value
-	gasMap["I16x8GtS"] = value
-	gasMap["I16x8GtU"] = value
-	gasMap["I16x8LeS"] = value
-	gasMap["I16x8LeU"] = value
-	gasMap["I16x8GeS"] = value
-	gasMap["I16x8GeU"] = value
-	gasMap["I32x4Eq"] = value
-	gasMap["I32x4Ne"] = value
-	gasMap["I32x4LtS"] = value
-	gasMap["I32x4LtU"] = value
-	gasMap["I32x4GtS"] = value
-	gasMap["I32x4GtU"] = value
-	gasMap["I32x4LeS"] = value
-	gasMap["I32x4LeU"] = value
-	gasMap["I32x4GeS"] = value
-	gasMap["I32x4GeU"] = value
-	gasMap["F32x4Eq"] = value
-	gasMap["F32x4Ne"] = value
-	gasMap["F32x4Lt"] = value
-	gasMap["F32x4Gt"] = value
-	gasMap["F32x4Le"] = value
-	gasMap["F32x4Ge"] = value
-	gasMap["F64x2Eq"] = value
-	gasMap["F64x2Ne"] = value
-	gasMap["F64x2Lt"] = value
-	gasMap["F64x2Gt"] = value
-	gasMap["F64x2Le"] = value
-	gasMap["F64x2Ge"] = value
-	gasMap["V128Not"] = value
-	gasMap["V128And"] = value
-	gasMap["V128Or"] = value
-	gasMap["V128Xor"] = value
-	gasMap["V128Bitselect"] = value
-	gasMap["I8x16Neg"] = value
-	gasMap["I8x16AnyTrue"] = value
-	gasMap["I8x16AllTrue"] = value
-	gasMap["I8x16Shl"] = value
-	gasMap["I8x16ShrS"] = value
-	gasMap["I8x16ShrU"] = value
-	gasMap["I8x16Add"] = value
-	gasMap["I8x16AddSaturateS"] = value
-	gasMap["I8x16AddSaturateU"] = value
-	gasMap["I8x16Sub"] = value
-	gasMap["I8x16SubSaturateS"] = value
-	gasMap["I8x16SubSaturateU"] = value
-	gasMap["I8x16Mul"] = value
-	gasMap["I16x8Neg"] = value
-	gasMap["I16x8AnyTrue"] = value
-	gasMap["I16x8AllTrue"] = value
-	gasMap["I16x8Shl"] = value
-	gasMap["I16x8ShrS"] = value
-	gasMap["I16x8ShrU"] = value
-	gasMap["I16x8Add"] = value
-	gasMap["I16x8AddSaturateS"] = value
-	gasMap["I16x8AddSaturateU"] = value
-	gasMap["I16x8Sub"] = value
-	gasMap["I16x8SubSaturateS"] = value
-	gasMap["I16x8SubSaturateU"] = value
-	gasMap["I16x8Mul"] = value
-	gasMap["I32x4Neg"] = value
-	gasMap["I32x4AnyTrue"] = value
-	gasMap["I32x4AllTrue"] = value
-	gasMap["I32x4Shl"] = value
-	gasMap["I32x4ShrS"] = value
-	gasMap["I32x4ShrU"] = value
-	gasMap["I32x4Add"] = value
-	gasMap["I32x4Sub"] = value
-	gasMap["I32x4Mul"] = value
-	gasMap["I64x2Neg"] = value
-	gasMap["I64x2AnyTrue"] = value
-	gasMap["I64x2AllTrue"] = value
-	gasMap["I64x2Shl"] = value
-	gasMap["I64x2ShrS"] = value
-	gasMap["I64x2ShrU"] = value
-	gasMap["I64x2Add"] = value
-	gasMap["I64x2Sub"] = value
-	gasMap["F32x4Abs"] = value
-	gasMap["F32x4Neg"] = value
-	gasMap["F32x4Sqrt"] = value
-	gasMap["F32x4Add"] = value
-	gasMap["F32x4Sub"] = value
-	gasMap["F32x4Mul"] = value
-	gasMap["F32x4Div"] = value
-	gasMap["F32x4Min"] = value
-	gasMap["F32x4Max"] = value
-	gasMap["F64x2Abs"] = value
-	gasMap["F64x2Neg"] = value
-	gasMap["F64x2Sqrt"] = value
-	gasMap["F64x2Add"] = value
-	gasMap["F64x2Sub"] = value
-	gasMap["F64x2Mul"] = value
-	gasMap["F64x2Div"] = value
-	gasMap["F64x2Min"] = value
-	gasMap["F64x2Max"] = value
-	gasMap["I32x4TruncSF32x4Sat"] = value
-	gasMap["I32x4TruncUF32x4Sat"] = value
-	gasMap["I64x2TruncSF64x2Sat"] = value
-	gasMap["I64x2TruncUF64x2Sat"] = value
-	gasMap["F32x4ConvertSI32x4"] = value
-	gasMap["F32x4ConvertUI32x4"] = value
-	gasMap["F64x2ConvertSI64x2"] = value
-	gasMap["F64x2ConvertUI64x2"] = value
-	gasMap["V8x16Swizzle"] = value
-	gasMap["V8x16Shuffle"] = value
-	gasMap["I8x16LoadSplat"] = value
-	gasMap["I16x8LoadSplat"] = value
-	gasMap["I32x4LoadSplat"] = value
-	gasMap["I64x2LoadSplat"] = value
-
-	return gasMap
-}
-
-func (opcode_costs_struct *WASMOpcodeCost) ToOpcodeCostsArray() [wasmer.OPCODE_COUNT]uint32 {
-	opcode_costs := [wasmer.OPCODE_COUNT]uint32{}
-
-	opcode_costs[wasmer.OpcodeUnreachable] = opcode_costs_struct.Unreachable
-	opcode_costs[wasmer.OpcodeNop] = opcode_costs_struct.Nop
-	opcode_costs[wasmer.OpcodeBlock] = opcode_costs_struct.Block
-	opcode_costs[wasmer.OpcodeLoop] = opcode_costs_struct.Loop
-	opcode_costs[wasmer.OpcodeIf] = opcode_costs_struct.If
-	opcode_costs[wasmer.OpcodeElse] = opcode_costs_struct.Else
-	opcode_costs[wasmer.OpcodeEnd] = opcode_costs_struct.End
-	opcode_costs[wasmer.OpcodeBr] = opcode_costs_struct.Br
-	opcode_costs[wasmer.OpcodeBrIf] = opcode_costs_struct.BrIf
-	opcode_costs[wasmer.OpcodeBrTable] = opcode_costs_struct.BrTable
-	opcode_costs[wasmer.OpcodeReturn] = opcode_costs_struct.Return
-	opcode_costs[wasmer.OpcodeCall] = opcode_costs_struct.Call
-	opcode_costs[wasmer.OpcodeCallIndirect] = opcode_costs_struct.CallIndirect
-	opcode_costs[wasmer.OpcodeDrop] = opcode_costs_struct.Drop
-	opcode_costs[wasmer.OpcodeSelect] = opcode_costs_struct.Select
-	opcode_costs[wasmer.OpcodeGetLocal] = opcode_costs_struct.GetLocal
-	opcode_costs[wasmer.OpcodeSetLocal] = opcode_costs_struct.SetLocal
-	opcode_costs[wasmer.OpcodeTeeLocal] = opcode_costs_struct.TeeLocal
-	opcode_costs[wasmer.OpcodeGetGlobal] = opcode_costs_struct.GetGlobal
-	opcode_costs[wasmer.OpcodeSetGlobal] = opcode_costs_struct.SetGlobal
-	opcode_costs[wasmer.OpcodeI32Load] = opcode_costs_struct.I32Load
-	opcode_costs[wasmer.OpcodeI64Load] = opcode_costs_struct.I64Load
-	opcode_costs[wasmer.OpcodeF32Load] = opcode_costs_struct.F32Load
-	opcode_costs[wasmer.OpcodeF64Load] = opcode_costs_struct.F64Load
-	opcode_costs[wasmer.OpcodeI32Load8S] = opcode_costs_struct.I32Load8S
-	opcode_costs[wasmer.OpcodeI32Load8U] = opcode_costs_struct.I32Load8U
-	opcode_costs[wasmer.OpcodeI32Load16S] = opcode_costs_struct.I32Load16S
-	opcode_costs[wasmer.OpcodeI32Load16U] = opcode_costs_struct.I32Load16U
-	opcode_costs[wasmer.OpcodeI64Load8S] = opcode_costs_struct.I64Load8S
-	opcode_costs[wasmer.OpcodeI64Load8U] = opcode_costs_struct.I64Load8U
-	opcode_costs[wasmer.OpcodeI64Load16S] = opcode_costs_struct.I64Load16S
-	opcode_costs[wasmer.OpcodeI64Load16U] = opcode_costs_struct.I64Load16U
-	opcode_costs[wasmer.OpcodeI64Load32S] = opcode_costs_struct.I64Load32S
-	opcode_costs[wasmer.OpcodeI64Load32U] = opcode_costs_struct.I64Load32U
-	opcode_costs[wasmer.OpcodeI32Store] = opcode_costs_struct.I32Store
-	opcode_costs[wasmer.OpcodeI64Store] = opcode_costs_struct.I64Store
-	opcode_costs[wasmer.OpcodeF32Store] = opcode_costs_struct.F32Store
-	opcode_costs[wasmer.OpcodeF64Store] = opcode_costs_struct.F64Store
-	opcode_costs[wasmer.OpcodeI32Store8] = opcode_costs_struct.I32Store8
-	opcode_costs[wasmer.OpcodeI32Store16] = opcode_costs_struct.I32Store16
-	opcode_costs[wasmer.OpcodeI64Store8] = opcode_costs_struct.I64Store8
-	opcode_costs[wasmer.OpcodeI64Store16] = opcode_costs_struct.I64Store16
-	opcode_costs[wasmer.OpcodeI64Store32] = opcode_costs_struct.I64Store32
-	opcode_costs[wasmer.OpcodeMemorySize] = opcode_costs_struct.MemorySize
-	opcode_costs[wasmer.OpcodeMemoryGrow] = opcode_costs_struct.MemoryGrow
-	opcode_costs[wasmer.OpcodeI32Const] = opcode_costs_struct.I32Const
-	opcode_costs[wasmer.OpcodeI64Const] = opcode_costs_struct.I64Const
-	opcode_costs[wasmer.OpcodeF32Const] = opcode_costs_struct.F32Const
-	opcode_costs[wasmer.OpcodeF64Const] = opcode_costs_struct.F64Const
-	opcode_costs[wasmer.OpcodeRefNull] = opcode_costs_struct.RefNull
-	opcode_costs[wasmer.OpcodeRefIsNull] = opcode_costs_struct.RefIsNull
-	opcode_costs[wasmer.OpcodeI32Eqz] = opcode_costs_struct.I32Eqz
-	opcode_costs[wasmer.OpcodeI32Eq] = opcode_costs_struct.I32Eq
-	opcode_costs[wasmer.OpcodeI32Ne] = opcode_costs_struct.I32Ne
-	opcode_costs[wasmer.OpcodeI32LtS] = opcode_costs_struct.I32LtS
-	opcode_costs[wasmer.OpcodeI32LtU] = opcode_costs_struct.I32LtU
-	opcode_costs[wasmer.OpcodeI32GtS] = opcode_costs_struct.I32GtS
-	opcode_costs[wasmer.OpcodeI32GtU] = opcode_costs_struct.I32GtU
-	opcode_costs[wasmer.OpcodeI32LeS] = opcode_costs_struct.I32LeS
-	opcode_costs[wasmer.OpcodeI32LeU] = opcode_costs_struct.I32LeU
-	opcode_costs[wasmer.OpcodeI32GeS] = opcode_costs_struct.I32GeS
-	opcode_costs[wasmer.OpcodeI32GeU] = opcode_costs_struct.I32GeU
-	opcode_costs[wasmer.OpcodeI64Eqz] = opcode_costs_struct.I64Eqz
-	opcode_costs[wasmer.OpcodeI64Eq] = opcode_costs_struct.I64Eq
-	opcode_costs[wasmer.OpcodeI64Ne] = opcode_costs_struct.I64Ne
-	opcode_costs[wasmer.OpcodeI64LtS] = opcode_costs_struct.I64LtS
-	opcode_costs[wasmer.OpcodeI64LtU] = opcode_costs_struct.I64LtU
-	opcode_costs[wasmer.OpcodeI64GtS] = opcode_costs_struct.I64GtS
-	opcode_costs[wasmer.OpcodeI64GtU] = opcode_costs_struct.I64GtU
-	opcode_costs[wasmer.OpcodeI64LeS] = opcode_costs_struct.I64LeS
-	opcode_costs[wasmer.OpcodeI64LeU] = opcode_costs_struct.I64LeU
-	opcode_costs[wasmer.OpcodeI64GeS] = opcode_costs_struct.I64GeS
-	opcode_costs[wasmer.OpcodeI64GeU] = opcode_costs_struct.I64GeU
-	opcode_costs[wasmer.OpcodeF32Eq] = opcode_costs_struct.F32Eq
-	opcode_costs[wasmer.OpcodeF32Ne] = opcode_costs_struct.F32Ne
-	opcode_costs[wasmer.OpcodeF32Lt] = opcode_costs_struct.F32Lt
-	opcode_costs[wasmer.OpcodeF32Gt] = opcode_costs_struct.F32Gt
-	opcode_costs[wasmer.OpcodeF32Le] = opcode_costs_struct.F32Le
-	opcode_costs[wasmer.OpcodeF32Ge] = opcode_costs_struct.F32Ge
-	opcode_costs[wasmer.OpcodeF64Eq] = opcode_costs_struct.F64Eq
-	opcode_costs[wasmer.OpcodeF64Ne] = opcode_costs_struct.F64Ne
-	opcode_costs[wasmer.OpcodeF64Lt] = opcode_costs_struct.F64Lt
-	opcode_costs[wasmer.OpcodeF64Gt] = opcode_costs_struct.F64Gt
-	opcode_costs[wasmer.OpcodeF64Le] = opcode_costs_struct.F64Le
-	opcode_costs[wasmer.OpcodeF64Ge] = opcode_costs_struct.F64Ge
-	opcode_costs[wasmer.OpcodeI32Clz] = opcode_costs_struct.I32Clz
-	opcode_costs[wasmer.OpcodeI32Ctz] = opcode_costs_struct.I32Ctz
-	opcode_costs[wasmer.OpcodeI32Popcnt] = opcode_costs_struct.I32Popcnt
-	opcode_costs[wasmer.OpcodeI32Add] = opcode_costs_struct.I32Add
-	opcode_costs[wasmer.OpcodeI32Sub] = opcode_costs_struct.I32Sub
-	opcode_costs[wasmer.OpcodeI32Mul] = opcode_costs_struct.I32Mul
-	opcode_costs[wasmer.OpcodeI32DivS] = opcode_costs_struct.I32DivS
-	opcode_costs[wasmer.OpcodeI32DivU] = opcode_costs_struct.I32DivU
-	opcode_costs[wasmer.OpcodeI32RemS] = opcode_costs_struct.I32RemS
-	opcode_costs[wasmer.OpcodeI32RemU] = opcode_costs_struct.I32RemU
-	opcode_costs[wasmer.OpcodeI32And] = opcode_costs_struct.I32And
-	opcode_costs[wasmer.OpcodeI32Or] = opcode_costs_struct.I32Or
-	opcode_costs[wasmer.OpcodeI32Xor] = opcode_costs_struct.I32Xor
-	opcode_costs[wasmer.OpcodeI32Shl] = opcode_costs_struct.I32Shl
-	opcode_costs[wasmer.OpcodeI32ShrS] = opcode_costs_struct.I32ShrS
-	opcode_costs[wasmer.OpcodeI32ShrU] = opcode_costs_struct.I32ShrU
-	opcode_costs[wasmer.OpcodeI32Rotl] = opcode_costs_struct.I32Rotl
-	opcode_costs[wasmer.OpcodeI32Rotr] = opcode_costs_struct.I32Rotr
-	opcode_costs[wasmer.OpcodeI64Clz] = opcode_costs_struct.I64Clz
-	opcode_costs[wasmer.OpcodeI64Ctz] = opcode_costs_struct.I64Ctz
-	opcode_costs[wasmer.OpcodeI64Popcnt] = opcode_costs_struct.I64Popcnt
-	opcode_costs[wasmer.OpcodeI64Add] = opcode_costs_struct.I64Add
-	opcode_costs[wasmer.OpcodeI64Sub] = opcode_costs_struct.I64Sub
-	opcode_costs[wasmer.OpcodeI64Mul] = opcode_costs_struct.I64Mul
-	opcode_costs[wasmer.OpcodeI64DivS] = opcode_costs_struct.I64DivS
-	opcode_costs[wasmer.OpcodeI64DivU] = opcode_costs_struct.I64DivU
-	opcode_costs[wasmer.OpcodeI64RemS] = opcode_costs_struct.I64RemS
-	opcode_costs[wasmer.OpcodeI64RemU] = opcode_costs_struct.I64RemU
-	opcode_costs[wasmer.OpcodeI64And] = opcode_costs_struct.I64And
-	opcode_costs[wasmer.OpcodeI64Or] = opcode_costs_struct.I64Or
-	opcode_costs[wasmer.OpcodeI64Xor] = opcode_costs_struct.I64Xor
-	opcode_costs[wasmer.OpcodeI64Shl] = opcode_costs_struct.I64Shl
-	opcode_costs[wasmer.OpcodeI64ShrS] = opcode_costs_struct.I64ShrS
-	opcode_costs[wasmer.OpcodeI64ShrU] = opcode_costs_struct.I64ShrU
-	opcode_costs[wasmer.OpcodeI64Rotl] = opcode_costs_struct.I64Rotl
-	opcode_costs[wasmer.OpcodeI64Rotr] = opcode_costs_struct.I64Rotr
-	opcode_costs[wasmer.OpcodeF32Abs] = opcode_costs_struct.F32Abs
-	opcode_costs[wasmer.OpcodeF32Neg] = opcode_costs_struct.F32Neg
-	opcode_costs[wasmer.OpcodeF32Ceil] = opcode_costs_struct.F32Ceil
-	opcode_costs[wasmer.OpcodeF32Floor] = opcode_costs_struct.F32Floor
-	opcode_costs[wasmer.OpcodeF32Trunc] = opcode_costs_struct.F32Trunc
-	opcode_costs[wasmer.OpcodeF32Nearest] = opcode_costs_struct.F32Nearest
-	opcode_costs[wasmer.OpcodeF32Sqrt] = opcode_costs_struct.F32Sqrt
-	opcode_costs[wasmer.OpcodeF32Add] = opcode_costs_struct.F32Add
-	opcode_costs[wasmer.OpcodeF32Sub] = opcode_costs_struct.F32Sub
-	opcode_costs[wasmer.OpcodeF32Mul] = opcode_costs_struct.F32Mul
-	opcode_costs[wasmer.OpcodeF32Div] = opcode_costs_struct.F32Div
-	opcode_costs[wasmer.OpcodeF32Min] = opcode_costs_struct.F32Min
-	opcode_costs[wasmer.OpcodeF32Max] = opcode_costs_struct.F32Max
-	opcode_costs[wasmer.OpcodeF32Copysign] = opcode_costs_struct.F32Copysign
-	opcode_costs[wasmer.OpcodeF64Abs] = opcode_costs_struct.F64Abs
-	opcode_costs[wasmer.OpcodeF64Neg] = opcode_costs_struct.F64Neg
-	opcode_costs[wasmer.OpcodeF64Ceil] = opcode_costs_struct.F64Ceil
-	opcode_costs[wasmer.OpcodeF64Floor] = opcode_costs_struct.F64Floor
-	opcode_costs[wasmer.OpcodeF64Trunc] = opcode_costs_struct.F64Trunc
-	opcode_costs[wasmer.OpcodeF64Nearest] = opcode_costs_struct.F64Nearest
-	opcode_costs[wasmer.OpcodeF64Sqrt] = opcode_costs_struct.F64Sqrt
-	opcode_costs[wasmer.OpcodeF64Add] = opcode_costs_struct.F64Add
-	opcode_costs[wasmer.OpcodeF64Sub] = opcode_costs_struct.F64Sub
-	opcode_costs[wasmer.OpcodeF64Mul] = opcode_costs_struct.F64Mul
-	opcode_costs[wasmer.OpcodeF64Div] = opcode_costs_struct.F64Div
-	opcode_costs[wasmer.OpcodeF64Min] = opcode_costs_struct.F64Min
-	opcode_costs[wasmer.OpcodeF64Max] = opcode_costs_struct.F64Max
-	opcode_costs[wasmer.OpcodeF64Copysign] = opcode_costs_struct.F64Copysign
-	opcode_costs[wasmer.OpcodeI32WrapI64] = opcode_costs_struct.I32WrapI64
-	opcode_costs[wasmer.OpcodeI32TruncSF32] = opcode_costs_struct.I32TruncSF32
-	opcode_costs[wasmer.OpcodeI32TruncUF32] = opcode_costs_struct.I32TruncUF32
-	opcode_costs[wasmer.OpcodeI32TruncSF64] = opcode_costs_struct.I32TruncSF64
-	opcode_costs[wasmer.OpcodeI32TruncUF64] = opcode_costs_struct.I32TruncUF64
-	opcode_costs[wasmer.OpcodeI64ExtendSI32] = opcode_costs_struct.I64ExtendSI32
-	opcode_costs[wasmer.OpcodeI64ExtendUI32] = opcode_costs_struct.I64ExtendUI32
-	opcode_costs[wasmer.OpcodeI64TruncSF32] = opcode_costs_struct.I64TruncSF32
-	opcode_costs[wasmer.OpcodeI64TruncUF32] = opcode_costs_struct.I64TruncUF32
-	opcode_costs[wasmer.OpcodeI64TruncSF64] = opcode_costs_struct.I64TruncSF64
-	opcode_costs[wasmer.OpcodeI64TruncUF64] = opcode_costs_struct.I64TruncUF64
-	opcode_costs[wasmer.OpcodeF32ConvertSI32] = opcode_costs_struct.F32ConvertSI32
-	opcode_costs[wasmer.OpcodeF32ConvertUI32] = opcode_costs_struct.F32ConvertUI32
-	opcode_costs[wasmer.OpcodeF32ConvertSI64] = opcode_costs_struct.F32ConvertSI64
-	opcode_costs[wasmer.OpcodeF32ConvertUI64] = opcode_costs_struct.F32ConvertUI64
-	opcode_costs[wasmer.OpcodeF32DemoteF64] = opcode_costs_struct.F32DemoteF64
-	opcode_costs[wasmer.OpcodeF64ConvertSI32] = opcode_costs_struct.F64ConvertSI32
-	opcode_costs[wasmer.OpcodeF64ConvertUI32] = opcode_costs_struct.F64ConvertUI32
-	opcode_costs[wasmer.OpcodeF64ConvertSI64] = opcode_costs_struct.F64ConvertSI64
-	opcode_costs[wasmer.OpcodeF64ConvertUI64] = opcode_costs_struct.F64ConvertUI64
-	opcode_costs[wasmer.OpcodeF64PromoteF32] = opcode_costs_struct.F64PromoteF32
-	opcode_costs[wasmer.OpcodeI32ReinterpretF32] = opcode_costs_struct.I32ReinterpretF32
-	opcode_costs[wasmer.OpcodeI64ReinterpretF64] = opcode_costs_struct.I64ReinterpretF64
-	opcode_costs[wasmer.OpcodeF32ReinterpretI32] = opcode_costs_struct.F32ReinterpretI32
-	opcode_costs[wasmer.OpcodeF64ReinterpretI64] = opcode_costs_struct.F64ReinterpretI64
-	opcode_costs[wasmer.OpcodeI32Extend8S] = opcode_costs_struct.I32Extend8S
-	opcode_costs[wasmer.OpcodeI32Extend16S] = opcode_costs_struct.I32Extend16S
-	opcode_costs[wasmer.OpcodeI64Extend8S] = opcode_costs_struct.I64Extend8S
-	opcode_costs[wasmer.OpcodeI64Extend16S] = opcode_costs_struct.I64Extend16S
-	opcode_costs[wasmer.OpcodeI64Extend32S] = opcode_costs_struct.I64Extend32S
-	opcode_costs[wasmer.OpcodeI32TruncSSatF32] = opcode_costs_struct.I32TruncSSatF32
-	opcode_costs[wasmer.OpcodeI32TruncUSatF32] = opcode_costs_struct.I32TruncUSatF32
-	opcode_costs[wasmer.OpcodeI32TruncSSatF64] = opcode_costs_struct.I32TruncSSatF64
-	opcode_costs[wasmer.OpcodeI32TruncUSatF64] = opcode_costs_struct.I32TruncUSatF64
-	opcode_costs[wasmer.OpcodeI64TruncSSatF32] = opcode_costs_struct.I64TruncSSatF32
-	opcode_costs[wasmer.OpcodeI64TruncUSatF32] = opcode_costs_struct.I64TruncUSatF32
-	opcode_costs[wasmer.OpcodeI64TruncSSatF64] = opcode_costs_struct.I64TruncSSatF64
-	opcode_costs[wasmer.OpcodeI64TruncUSatF64] = opcode_costs_struct.I64TruncUSatF64
-	opcode_costs[wasmer.OpcodeMemoryInit] = opcode_costs_struct.MemoryInit
-	opcode_costs[wasmer.OpcodeDataDrop] = opcode_costs_struct.DataDrop
-	opcode_costs[wasmer.OpcodeMemoryCopy] = opcode_costs_struct.MemoryCopy
-	opcode_costs[wasmer.OpcodeMemoryFill] = opcode_costs_struct.MemoryFill
-	opcode_costs[wasmer.OpcodeTableInit] = opcode_costs_struct.TableInit
-	opcode_costs[wasmer.OpcodeElemDrop] = opcode_costs_struct.ElemDrop
-	opcode_costs[wasmer.OpcodeTableCopy] = opcode_costs_struct.TableCopy
-	opcode_costs[wasmer.OpcodeTableGet] = opcode_costs_struct.TableGet
-	opcode_costs[wasmer.OpcodeTableSet] = opcode_costs_struct.TableSet
-	opcode_costs[wasmer.OpcodeTableGrow] = opcode_costs_struct.TableGrow
-	opcode_costs[wasmer.OpcodeTableSize] = opcode_costs_struct.TableSize
-	opcode_costs[wasmer.OpcodeWake] = opcode_costs_struct.Wake
-	opcode_costs[wasmer.OpcodeI32Wait] = opcode_costs_struct.I32Wait
-	opcode_costs[wasmer.OpcodeI64Wait] = opcode_costs_struct.I64Wait
-	opcode_costs[wasmer.OpcodeFence] = opcode_costs_struct.Fence
-	opcode_costs[wasmer.OpcodeI32AtomicLoad] = opcode_costs_struct.I32AtomicLoad
-	opcode_costs[wasmer.OpcodeI64AtomicLoad] = opcode_costs_struct.I64AtomicLoad
-	opcode_costs[wasmer.OpcodeI32AtomicLoad8U] = opcode_costs_struct.I32AtomicLoad8U
-	opcode_costs[wasmer.OpcodeI32AtomicLoad16U] = opcode_costs_struct.I32AtomicLoad16U
-	opcode_costs[wasmer.OpcodeI64AtomicLoad8U] = opcode_costs_struct.I64AtomicLoad8U
-	opcode_costs[wasmer.OpcodeI64AtomicLoad16U] = opcode_costs_struct.I64AtomicLoad16U
-	opcode_costs[wasmer.OpcodeI64AtomicLoad32U] = opcode_costs_struct.I64AtomicLoad32U
-	opcode_costs[wasmer.OpcodeI32AtomicStore] = opcode_costs_struct.I32AtomicStore
-	opcode_costs[wasmer.OpcodeI64AtomicStore] = opcode_costs_struct.I64AtomicStore
-	opcode_costs[wasmer.OpcodeI32AtomicStore8] = opcode_costs_struct.I32AtomicStore8
-	opcode_costs[wasmer.OpcodeI32AtomicStore16] = opcode_costs_struct.I32AtomicStore16
-	opcode_costs[wasmer.OpcodeI64AtomicStore8] = opcode_costs_struct.I64AtomicStore8
-	opcode_costs[wasmer.OpcodeI64AtomicStore16] = opcode_costs_struct.I64AtomicStore16
-	opcode_costs[wasmer.OpcodeI64AtomicStore32] = opcode_costs_struct.I64AtomicStore32
-	opcode_costs[wasmer.OpcodeI32AtomicRmwAdd] = opcode_costs_struct.I32AtomicRmwAdd
-	opcode_costs[wasmer.OpcodeI64AtomicRmwAdd] = opcode_costs_struct.I64AtomicRmwAdd
-	opcode_costs[wasmer.OpcodeI32AtomicRmw8UAdd] = opcode_costs_struct.I32AtomicRmw8UAdd
-	opcode_costs[wasmer.OpcodeI32AtomicRmw16UAdd] = opcode_costs_struct.I32AtomicRmw16UAdd
-	opcode_costs[wasmer.OpcodeI64AtomicRmw8UAdd] = opcode_costs_struct.I64AtomicRmw8UAdd
-	opcode_costs[wasmer.OpcodeI64AtomicRmw16UAdd] = opcode_costs_struct.I64AtomicRmw16UAdd
-	opcode_costs[wasmer.OpcodeI64AtomicRmw32UAdd] = opcode_costs_struct.I64AtomicRmw32UAdd
-	opcode_costs[wasmer.OpcodeI32AtomicRmwSub] = opcode_costs_struct.I32AtomicRmwSub
-	opcode_costs[wasmer.OpcodeI64AtomicRmwSub] = opcode_costs_struct.I64AtomicRmwSub
-	opcode_costs[wasmer.OpcodeI32AtomicRmw8USub] = opcode_costs_struct.I32AtomicRmw8USub
-	opcode_costs[wasmer.OpcodeI32AtomicRmw16USub] = opcode_costs_struct.I32AtomicRmw16USub
-	opcode_costs[wasmer.OpcodeI64AtomicRmw8USub] = opcode_costs_struct.I64AtomicRmw8USub
-	opcode_costs[wasmer.OpcodeI64AtomicRmw16USub] = opcode_costs_struct.I64AtomicRmw16USub
-	opcode_costs[wasmer.OpcodeI64AtomicRmw32USub] = opcode_costs_struct.I64AtomicRmw32USub
-	opcode_costs[wasmer.OpcodeI32AtomicRmwAnd] = opcode_costs_struct.I32AtomicRmwAnd
-	opcode_costs[wasmer.OpcodeI64AtomicRmwAnd] = opcode_costs_struct.I64AtomicRmwAnd
-	opcode_costs[wasmer.OpcodeI32AtomicRmw8UAnd] = opcode_costs_struct.I32AtomicRmw8UAnd
-	opcode_costs[wasmer.OpcodeI32AtomicRmw16UAnd] = opcode_costs_struct.I32AtomicRmw16UAnd
-	opcode_costs[wasmer.OpcodeI64AtomicRmw8UAnd] = opcode_costs_struct.I64AtomicRmw8UAnd
-	opcode_costs[wasmer.OpcodeI64AtomicRmw16UAnd] = opcode_costs_struct.I64AtomicRmw16UAnd
-	opcode_costs[wasmer.OpcodeI64AtomicRmw32UAnd] = opcode_costs_struct.I64AtomicRmw32UAnd
-	opcode_costs[wasmer.OpcodeI32AtomicRmwOr] = opcode_costs_struct.I32AtomicRmwOr
-	opcode_costs[wasmer.OpcodeI64AtomicRmwOr] = opcode_costs_struct.I64AtomicRmwOr
-	opcode_costs[wasmer.OpcodeI32AtomicRmw8UOr] = opcode_costs_struct.I32AtomicRmw8UOr
-	opcode_costs[wasmer.OpcodeI32AtomicRmw16UOr] = opcode_costs_struct.I32AtomicRmw16UOr
-	opcode_costs[wasmer.OpcodeI64AtomicRmw8UOr] = opcode_costs_struct.I64AtomicRmw8UOr
-	opcode_costs[wasmer.OpcodeI64AtomicRmw16UOr] = opcode_costs_struct.I64AtomicRmw16UOr
-	opcode_costs[wasmer.OpcodeI64AtomicRmw32UOr] = opcode_costs_struct.I64AtomicRmw32UOr
-	opcode_costs[wasmer.OpcodeI32AtomicRmwXor] = opcode_costs_struct.I32AtomicRmwXor
-	opcode_costs[wasmer.OpcodeI64AtomicRmwXor] = opcode_costs_struct.I64AtomicRmwXor
-	opcode_costs[wasmer.OpcodeI32AtomicRmw8UXor] = opcode_costs_struct.I32AtomicRmw8UXor
-	opcode_costs[wasmer.OpcodeI32AtomicRmw16UXor] = opcode_costs_struct.I32AtomicRmw16UXor
-	opcode_costs[wasmer.OpcodeI64AtomicRmw8UXor] = opcode_costs_struct.I64AtomicRmw8UXor
-	opcode_costs[wasmer.OpcodeI64AtomicRmw16UXor] = opcode_costs_struct.I64AtomicRmw16UXor
-	opcode_costs[wasmer.OpcodeI64AtomicRmw32UXor] = opcode_costs_struct.I64AtomicRmw32UXor
-	opcode_costs[wasmer.OpcodeI32AtomicRmwXchg] = opcode_costs_struct.I32AtomicRmwXchg
-	opcode_costs[wasmer.OpcodeI64AtomicRmwXchg] = opcode_costs_struct.I64AtomicRmwXchg
-	opcode_costs[wasmer.OpcodeI32AtomicRmw8UXchg] = opcode_costs_struct.I32AtomicRmw8UXchg
-	opcode_costs[wasmer.OpcodeI32AtomicRmw16UXchg] = opcode_costs_struct.I32AtomicRmw16UXchg
-	opcode_costs[wasmer.OpcodeI64AtomicRmw8UXchg] = opcode_costs_struct.I64AtomicRmw8UXchg
-	opcode_costs[wasmer.OpcodeI64AtomicRmw16UXchg] = opcode_costs_struct.I64AtomicRmw16UXchg
-	opcode_costs[wasmer.OpcodeI64AtomicRmw32UXchg] = opcode_costs_struct.I64AtomicRmw32UXchg
-	opcode_costs[wasmer.OpcodeI32AtomicRmwCmpxchg] = opcode_costs_struct.I32AtomicRmwCmpxchg
-	opcode_costs[wasmer.OpcodeI64AtomicRmwCmpxchg] = opcode_costs_struct.I64AtomicRmwCmpxchg
-	opcode_costs[wasmer.OpcodeI32AtomicRmw8UCmpxchg] = opcode_costs_struct.I32AtomicRmw8UCmpxchg
-	opcode_costs[wasmer.OpcodeI32AtomicRmw16UCmpxchg] = opcode_costs_struct.I32AtomicRmw16UCmpxchg
-	opcode_costs[wasmer.OpcodeI64AtomicRmw8UCmpxchg] = opcode_costs_struct.I64AtomicRmw8UCmpxchg
-	opcode_costs[wasmer.OpcodeI64AtomicRmw16UCmpxchg] = opcode_costs_struct.I64AtomicRmw16UCmpxchg
-	opcode_costs[wasmer.OpcodeI64AtomicRmw32UCmpxchg] = opcode_costs_struct.I64AtomicRmw32UCmpxchg
-	opcode_costs[wasmer.OpcodeV128Load] = opcode_costs_struct.V128Load
-	opcode_costs[wasmer.OpcodeV128Store] = opcode_costs_struct.V128Store
-	opcode_costs[wasmer.OpcodeV128Const] = opcode_costs_struct.V128Const
-	opcode_costs[wasmer.OpcodeI8x16Splat] = opcode_costs_struct.I8x16Splat
-	opcode_costs[wasmer.OpcodeI8x16ExtractLaneS] = opcode_costs_struct.I8x16ExtractLaneS
-	opcode_costs[wasmer.OpcodeI8x16ExtractLaneU] = opcode_costs_struct.I8x16ExtractLaneU
-	opcode_costs[wasmer.OpcodeI8x16ReplaceLane] = opcode_costs_struct.I8x16ReplaceLane
-	opcode_costs[wasmer.OpcodeI16x8Splat] = opcode_costs_struct.I16x8Splat
-	opcode_costs[wasmer.OpcodeI16x8ExtractLaneS] = opcode_costs_struct.I16x8ExtractLaneS
-	opcode_costs[wasmer.OpcodeI16x8ExtractLaneU] = opcode_costs_struct.I16x8ExtractLaneU
-	opcode_costs[wasmer.OpcodeI16x8ReplaceLane] = opcode_costs_struct.I16x8ReplaceLane
-	opcode_costs[wasmer.OpcodeI32x4Splat] = opcode_costs_struct.I32x4Splat
-	opcode_costs[wasmer.OpcodeI32x4ExtractLane] = opcode_costs_struct.I32x4ExtractLane
-	opcode_costs[wasmer.OpcodeI32x4ReplaceLane] = opcode_costs_struct.I32x4ReplaceLane
-	opcode_costs[wasmer.OpcodeI64x2Splat] = opcode_costs_struct.I64x2Splat
-	opcode_costs[wasmer.OpcodeI64x2ExtractLane] = opcode_costs_struct.I64x2ExtractLane
-	opcode_costs[wasmer.OpcodeI64x2ReplaceLane] = opcode_costs_struct.I64x2ReplaceLane
-	opcode_costs[wasmer.OpcodeF32x4Splat] = opcode_costs_struct.F32x4Splat
-	opcode_costs[wasmer.OpcodeF32x4ExtractLane] = opcode_costs_struct.F32x4ExtractLane
-	opcode_costs[wasmer.OpcodeF32x4ReplaceLane] = opcode_costs_struct.F32x4ReplaceLane
-	opcode_costs[wasmer.OpcodeF64x2Splat] = opcode_costs_struct.F64x2Splat
-	opcode_costs[wasmer.OpcodeF64x2ExtractLane] = opcode_costs_struct.F64x2ExtractLane
-	opcode_costs[wasmer.OpcodeF64x2ReplaceLane] = opcode_costs_struct.F64x2ReplaceLane
-	opcode_costs[wasmer.OpcodeI8x16Eq] = opcode_costs_struct.I8x16Eq
-	opcode_costs[wasmer.OpcodeI8x16Ne] = opcode_costs_struct.I8x16Ne
-	opcode_costs[wasmer.OpcodeI8x16LtS] = opcode_costs_struct.I8x16LtS
-	opcode_costs[wasmer.OpcodeI8x16LtU] = opcode_costs_struct.I8x16LtU
-	opcode_costs[wasmer.OpcodeI8x16GtS] = opcode_costs_struct.I8x16GtS
-	opcode_costs[wasmer.OpcodeI8x16GtU] = opcode_costs_struct.I8x16GtU
-	opcode_costs[wasmer.OpcodeI8x16LeS] = opcode_costs_struct.I8x16LeS
-	opcode_costs[wasmer.OpcodeI8x16LeU] = opcode_costs_struct.I8x16LeU
-	opcode_costs[wasmer.OpcodeI8x16GeS] = opcode_costs_struct.I8x16GeS
-	opcode_costs[wasmer.OpcodeI8x16GeU] = opcode_costs_struct.I8x16GeU
-	opcode_costs[wasmer.OpcodeI16x8Eq] = opcode_costs_struct.I16x8Eq
-	opcode_costs[wasmer.OpcodeI16x8Ne] = opcode_costs_struct.I16x8Ne
-	opcode_costs[wasmer.OpcodeI16x8LtS] = opcode_costs_struct.I16x8LtS
-	opcode_costs[wasmer.OpcodeI16x8LtU] = opcode_costs_struct.I16x8LtU
-	opcode_costs[wasmer.OpcodeI16x8GtS] = opcode_costs_struct.I16x8GtS
-	opcode_costs[wasmer.OpcodeI16x8GtU] = opcode_costs_struct.I16x8GtU
-	opcode_costs[wasmer.OpcodeI16x8LeS] = opcode_costs_struct.I16x8LeS
-	opcode_costs[wasmer.OpcodeI16x8LeU] = opcode_costs_struct.I16x8LeU
-	opcode_costs[wasmer.OpcodeI16x8GeS] = opcode_costs_struct.I16x8GeS
-	opcode_costs[wasmer.OpcodeI16x8GeU] = opcode_costs_struct.I16x8GeU
-	opcode_costs[wasmer.OpcodeI32x4Eq] = opcode_costs_struct.I32x4Eq
-	opcode_costs[wasmer.OpcodeI32x4Ne] = opcode_costs_struct.I32x4Ne
-	opcode_costs[wasmer.OpcodeI32x4LtS] = opcode_costs_struct.I32x4LtS
-	opcode_costs[wasmer.OpcodeI32x4LtU] = opcode_costs_struct.I32x4LtU
-	opcode_costs[wasmer.OpcodeI32x4GtS] = opcode_costs_struct.I32x4GtS
-	opcode_costs[wasmer.OpcodeI32x4GtU] = opcode_costs_struct.I32x4GtU
-	opcode_costs[wasmer.OpcodeI32x4LeS] = opcode_costs_struct.I32x4LeS
-	opcode_costs[wasmer.OpcodeI32x4LeU] = opcode_costs_struct.I32x4LeU
-	opcode_costs[wasmer.OpcodeI32x4GeS] = opcode_costs_struct.I32x4GeS
-	opcode_costs[wasmer.OpcodeI32x4GeU] = opcode_costs_struct.I32x4GeU
-	opcode_costs[wasmer.OpcodeF32x4Eq] = opcode_costs_struct.F32x4Eq
-	opcode_costs[wasmer.OpcodeF32x4Ne] = opcode_costs_struct.F32x4Ne
-	opcode_costs[wasmer.OpcodeF32x4Lt] = opcode_costs_struct.F32x4Lt
-	opcode_costs[wasmer.OpcodeF32x4Gt] = opcode_costs_struct.F32x4Gt
-	opcode_costs[wasmer.OpcodeF32x4Le] = opcode_costs_struct.F32x4Le
-	opcode_costs[wasmer.OpcodeF32x4Ge] = opcode_costs_struct.F32x4Ge
-	opcode_costs[wasmer.OpcodeF64x2Eq] = opcode_costs_struct.F64x2Eq
-	opcode_costs[wasmer.OpcodeF64x2Ne] = opcode_costs_struct.F64x2Ne
-	opcode_costs[wasmer.OpcodeF64x2Lt] = opcode_costs_struct.F64x2Lt
-	opcode_costs[wasmer.OpcodeF64x2Gt] = opcode_costs_struct.F64x2Gt
-	opcode_costs[wasmer.OpcodeF64x2Le] = opcode_costs_struct.F64x2Le
-	opcode_costs[wasmer.OpcodeF64x2Ge] = opcode_costs_struct.F64x2Ge
-	opcode_costs[wasmer.OpcodeV128Not] = opcode_costs_struct.V128Not
-	opcode_costs[wasmer.OpcodeV128And] = opcode_costs_struct.V128And
-	opcode_costs[wasmer.OpcodeV128Or] = opcode_costs_struct.V128Or
-	opcode_costs[wasmer.OpcodeV128Xor] = opcode_costs_struct.V128Xor
-	opcode_costs[wasmer.OpcodeV128Bitselect] = opcode_costs_struct.V128Bitselect
-	opcode_costs[wasmer.OpcodeI8x16Neg] = opcode_costs_struct.I8x16Neg
-	opcode_costs[wasmer.OpcodeI8x16AnyTrue] = opcode_costs_struct.I8x16AnyTrue
-	opcode_costs[wasmer.OpcodeI8x16AllTrue] = opcode_costs_struct.I8x16AllTrue
-	opcode_costs[wasmer.OpcodeI8x16Shl] = opcode_costs_struct.I8x16Shl
-	opcode_costs[wasmer.OpcodeI8x16ShrS] = opcode_costs_struct.I8x16ShrS
-	opcode_costs[wasmer.OpcodeI8x16ShrU] = opcode_costs_struct.I8x16ShrU
-	opcode_costs[wasmer.OpcodeI8x16Add] = opcode_costs_struct.I8x16Add
-	opcode_costs[wasmer.OpcodeI8x16AddSaturateS] = opcode_costs_struct.I8x16AddSaturateS
-	opcode_costs[wasmer.OpcodeI8x16AddSaturateU] = opcode_costs_struct.I8x16AddSaturateU
-	opcode_costs[wasmer.OpcodeI8x16Sub] = opcode_costs_struct.I8x16Sub
-	opcode_costs[wasmer.OpcodeI8x16SubSaturateS] = opcode_costs_struct.I8x16SubSaturateS
-	opcode_costs[wasmer.OpcodeI8x16SubSaturateU] = opcode_costs_struct.I8x16SubSaturateU
-	opcode_costs[wasmer.OpcodeI8x16Mul] = opcode_costs_struct.I8x16Mul
-	opcode_costs[wasmer.OpcodeI16x8Neg] = opcode_costs_struct.I16x8Neg
-	opcode_costs[wasmer.OpcodeI16x8AnyTrue] = opcode_costs_struct.I16x8AnyTrue
-	opcode_costs[wasmer.OpcodeI16x8AllTrue] = opcode_costs_struct.I16x8AllTrue
-	opcode_costs[wasmer.OpcodeI16x8Shl] = opcode_costs_struct.I16x8Shl
-	opcode_costs[wasmer.OpcodeI16x8ShrS] = opcode_costs_struct.I16x8ShrS
-	opcode_costs[wasmer.OpcodeI16x8ShrU] = opcode_costs_struct.I16x8ShrU
-	opcode_costs[wasmer.OpcodeI16x8Add] = opcode_costs_struct.I16x8Add
-	opcode_costs[wasmer.OpcodeI16x8AddSaturateS] = opcode_costs_struct.I16x8AddSaturateS
-	opcode_costs[wasmer.OpcodeI16x8AddSaturateU] = opcode_costs_struct.I16x8AddSaturateU
-	opcode_costs[wasmer.OpcodeI16x8Sub] = opcode_costs_struct.I16x8Sub
-	opcode_costs[wasmer.OpcodeI16x8SubSaturateS] = opcode_costs_struct.I16x8SubSaturateS
-	opcode_costs[wasmer.OpcodeI16x8SubSaturateU] = opcode_costs_struct.I16x8SubSaturateU
-	opcode_costs[wasmer.OpcodeI16x8Mul] = opcode_costs_struct.I16x8Mul
-	opcode_costs[wasmer.OpcodeI32x4Neg] = opcode_costs_struct.I32x4Neg
-	opcode_costs[wasmer.OpcodeI32x4AnyTrue] = opcode_costs_struct.I32x4AnyTrue
-	opcode_costs[wasmer.OpcodeI32x4AllTrue] = opcode_costs_struct.I32x4AllTrue
-	opcode_costs[wasmer.OpcodeI32x4Shl] = opcode_costs_struct.I32x4Shl
-	opcode_costs[wasmer.OpcodeI32x4ShrS] = opcode_costs_struct.I32x4ShrS
-	opcode_costs[wasmer.OpcodeI32x4ShrU] = opcode_costs_struct.I32x4ShrU
-	opcode_costs[wasmer.OpcodeI32x4Add] = opcode_costs_struct.I32x4Add
-	opcode_costs[wasmer.OpcodeI32x4Sub] = opcode_costs_struct.I32x4Sub
-	opcode_costs[wasmer.OpcodeI32x4Mul] = opcode_costs_struct.I32x4Mul
-	opcode_costs[wasmer.OpcodeI64x2Neg] = opcode_costs_struct.I64x2Neg
-	opcode_costs[wasmer.OpcodeI64x2AnyTrue] = opcode_costs_struct.I64x2AnyTrue
-	opcode_costs[wasmer.OpcodeI64x2AllTrue] = opcode_costs_struct.I64x2AllTrue
-	opcode_costs[wasmer.OpcodeI64x2Shl] = opcode_costs_struct.I64x2Shl
-	opcode_costs[wasmer.OpcodeI64x2ShrS] = opcode_costs_struct.I64x2ShrS
-	opcode_costs[wasmer.OpcodeI64x2ShrU] = opcode_costs_struct.I64x2ShrU
-	opcode_costs[wasmer.OpcodeI64x2Add] = opcode_costs_struct.I64x2Add
-	opcode_costs[wasmer.OpcodeI64x2Sub] = opcode_costs_struct.I64x2Sub
-	opcode_costs[wasmer.OpcodeF32x4Abs] = opcode_costs_struct.F32x4Abs
-	opcode_costs[wasmer.OpcodeF32x4Neg] = opcode_costs_struct.F32x4Neg
-	opcode_costs[wasmer.OpcodeF32x4Sqrt] = opcode_costs_struct.F32x4Sqrt
-	opcode_costs[wasmer.OpcodeF32x4Add] = opcode_costs_struct.F32x4Add
-	opcode_costs[wasmer.OpcodeF32x4Sub] = opcode_costs_struct.F32x4Sub
-	opcode_costs[wasmer.OpcodeF32x4Mul] = opcode_costs_struct.F32x4Mul
-	opcode_costs[wasmer.OpcodeF32x4Div] = opcode_costs_struct.F32x4Div
-	opcode_costs[wasmer.OpcodeF32x4Min] = opcode_costs_struct.F32x4Min
-	opcode_costs[wasmer.OpcodeF32x4Max] = opcode_costs_struct.F32x4Max
-	opcode_costs[wasmer.OpcodeF64x2Abs] = opcode_costs_struct.F64x2Abs
-	opcode_costs[wasmer.OpcodeF64x2Neg] = opcode_costs_struct.F64x2Neg
-	opcode_costs[wasmer.OpcodeF64x2Sqrt] = opcode_costs_struct.F64x2Sqrt
-	opcode_costs[wasmer.OpcodeF64x2Add] = opcode_costs_struct.F64x2Add
-	opcode_costs[wasmer.OpcodeF64x2Sub] = opcode_costs_struct.F64x2Sub
-	opcode_costs[wasmer.OpcodeF64x2Mul] = opcode_costs_struct.F64x2Mul
-	opcode_costs[wasmer.OpcodeF64x2Div] = opcode_costs_struct.F64x2Div
-	opcode_costs[wasmer.OpcodeF64x2Min] = opcode_costs_struct.F64x2Min
-	opcode_costs[wasmer.OpcodeF64x2Max] = opcode_costs_struct.F64x2Max
-	opcode_costs[wasmer.OpcodeI32x4TruncSF32x4Sat] = opcode_costs_struct.I32x4TruncSF32x4Sat
-	opcode_costs[wasmer.OpcodeI32x4TruncUF32x4Sat] = opcode_costs_struct.I32x4TruncUF32x4Sat
-	opcode_costs[wasmer.OpcodeI64x2TruncSF64x2Sat] = opcode_costs_struct.I64x2TruncSF64x2Sat
-	opcode_costs[wasmer.OpcodeI64x2TruncUF64x2Sat] = opcode_costs_struct.I64x2TruncUF64x2Sat
-	opcode_costs[wasmer.OpcodeF32x4ConvertSI32x4] = opcode_costs_struct.F32x4ConvertSI32x4
-	opcode_costs[wasmer.OpcodeF32x4ConvertUI32x4] = opcode_costs_struct.F32x4ConvertUI32x4
-	opcode_costs[wasmer.OpcodeF64x2ConvertSI64x2] = opcode_costs_struct.F64x2ConvertSI64x2
-	opcode_costs[wasmer.OpcodeF64x2ConvertUI64x2] = opcode_costs_struct.F64x2ConvertUI64x2
-	opcode_costs[wasmer.OpcodeV8x16Swizzle] = opcode_costs_struct.V8x16Swizzle
-	opcode_costs[wasmer.OpcodeV8x16Shuffle] = opcode_costs_struct.V8x16Shuffle
-	opcode_costs[wasmer.OpcodeI8x16LoadSplat] = opcode_costs_struct.I8x16LoadSplat
-	opcode_costs[wasmer.OpcodeI16x8LoadSplat] = opcode_costs_struct.I16x8LoadSplat
-	opcode_costs[wasmer.OpcodeI32x4LoadSplat] = opcode_costs_struct.I32x4LoadSplat
-	opcode_costs[wasmer.OpcodeI64x2LoadSplat] = opcode_costs_struct.I64x2LoadSplat
-
-	return opcode_costs
-}
